+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/CloudToolsPluginInitializationComponent.java
@@ -16,10 +16,10 @@
 
 package com.google.cloud.tools.intellij;
 
+import com.google.cloud.tools.intellij.appengine.action.AppEngineToolsMenuAction;
+import com.google.cloud.tools.intellij.appengine.action.configuration.MavenBuildDeploymentSourceType;
+import com.google.cloud.tools.intellij.appengine.action.configuration.UserSpecifiedPathDeploymentSourceType;
 import com.google.cloud.tools.intellij.appengine.cloud.AppEngineCloudType;
-import com.google.cloud.tools.intellij.appengine.cloud.AppEngineToolsMenuAction;
-import com.google.cloud.tools.intellij.appengine.cloud.MavenBuildDeploymentSourceType;
-import com.google.cloud.tools.intellij.appengine.cloud.UserSpecifiedPathDeploymentSourceType;
 import com.google.cloud.tools.intellij.debugger.CloudDebugConfigType;
 
 import com.intellij.execution.configurations.ConfigurationType;
