+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/appengine/action/AppEngineAction.java
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.appengine.cloud;
+package com.google.cloud.tools.intellij.appengine.action;
 
 import com.google.cloud.tools.app.api.AppEngineException;
 import com.google.cloud.tools.app.impl.cloudsdk.internal.process.DefaultProcessRunner;
 import com.google.cloud.tools.app.impl.cloudsdk.internal.sdk.CloudSdk;
+import com.google.cloud.tools.intellij.appengine.action.configuration.AppEngineDeploymentConfiguration;
 import com.google.cloud.tools.intellij.login.CredentialedUser;
 import com.google.cloud.tools.intellij.login.Services;
 import com.google.cloud.tools.intellij.util.GctBundle;
@@ -102,7 +103,7 @@ public abstract class AppEngineAction implements Runnable {
   /**
    * Kill any executing process for the action.
    */
-  protected void cancel() {
+  public void cancel() {
     if (processRunner != null
         && processRunner.getProcess() != null) {
       cancelled = true;
