+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/appengine/action/configuration/AppEngineDeploymentConfigurator.java
@@ -14,7 +14,11 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.appengine.cloud;
+package com.google.cloud.tools.intellij.appengine.action.configuration;
+
+import com.google.cloud.tools.intellij.appengine.action.CloudSdkAppEngineHelper;
+import com.google.cloud.tools.intellij.appengine.action.ui.AppEngineDeploymentRunConfigurationEditor;
+import com.google.cloud.tools.intellij.appengine.cloud.AppEngineServerConfiguration;
 
 import com.intellij.openapi.module.Module;
 import com.intellij.openapi.module.ModuleManager;
@@ -42,7 +46,7 @@ import java.util.List;
 /**
  * Sets up the configuration elements for an AppEngine Cloud deployment.
  */
-class AppEngineDeploymentConfigurator extends
+public class AppEngineDeploymentConfigurator extends
     DeploymentConfigurator<AppEngineDeploymentConfiguration, AppEngineServerConfiguration> {
 
   private final Project project;
