+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/appengine/action/configuration/MavenBuildDeploymentSourceType.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.appengine.cloud;
+package com.google.cloud.tools.intellij.appengine.action.configuration;
 
 import com.google.common.base.Predicate;
 import com.google.common.collect.Collections2;
