+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/appengine/action/ui/AppEngineDeploymentRunConfigurationEditor.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.google.cloud.tools.intellij.appengine.cloud.AppEngineDeploymentRunConfigurationEditor">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.google.cloud.tools.intellij.appengine.action.ui.AppEngineDeploymentRunConfigurationEditor">
   <grid id="1c5af" binding="editorPanel" layout-manager="GridLayoutManager" row-count="3" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="8" left="0" bottom="0" right="0"/>
     <constraints>
