+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/appengine/action/ui/SelectConfigDestinationFolderDialog.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.appengine.cloud;
+package com.google.cloud.tools.intellij.appengine.action.ui;
 
 import com.google.cloud.tools.intellij.util.GctBundle;
